#48: moving retrieve JWKs functions to `jwks.go` file
